feat(ui): split dnd overlay to separate component
This reduces top-level rerenders when zooming in and out on workflow editor
